+++ src/commands/apply.js
@@ -36,6 +36,7 @@ ApplyCommand.flags = {
 
 ApplyCommand.examples = [
   '$ sshpm apply openssh',
+  '$ sshpm apply putty',
   '$ sshpm apply winscp',
   '$ sshpm apply winscp-portable -c winscp.ini',
   '$ sshpm apply zoc'