+++ src/exporters/putty.js
@@ -1,0 +1,25 @@
+module.exports = async (cli, data, flags) => {
+  const { Registry } = require('rage-edit')
+  const childProcess = require('child_process')
+
+  if (!flags.keep) {
+    const conf = await Registry.get('HKCU\\SOFTWARE\\SimonTatham\\PuTTY\\Sessions')
+    for (const element of Object.keys(conf)) {
+      const x = await Registry.get('HKCU\\SOFTWARE\\SimonTatham\\PuTTY\\Sessions\\' + element)
+      if (!x || !x.$values.sshpm) continue
+      await Registry.delete('HKCU\\SOFTWARE\\SimonTatham\\PuTTY\\Sessions\\' + element)
+    }
+  }
+
+  data.forEach(async element => {
+    if (element.key && !flags.conf) return cli.log('Error: WinSCP.com not specified, please point to it using -c')
+
+    if (element.key) childProcess.spawn(flags.conf, ['/keygen', element.key, '/output=' + element.key + '.ppk'])
+    const key = 'HKCU\\SOFTWARE\\SimonTatham\\PuTTY\\Sessions\\' + element.name.replace(/\s/g, '%20')
+    await Registry.set(key, 'HostName', element.host)
+    await Registry.set(key, 'PortNumber', Number(element.port))
+    await Registry.set(key, 'UserName', element.user)
+    if (element.key) await Registry.set(key, 'PublicKeyFile', element.key + '.ppk')
+    await Registry.set(key, 'Sshpm', cli.config.version)
+  })
+}