+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Unix-like aliases (touch, ls, rm)
+- Documentation for installation with PNPM
+- Documentation for installation with Yarn
 
 ## [0.2.0] - 2020-08-05