+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Documentation for installation with PNPM
 - Documentation for installation with Yarn
 
+### Changed
+
+- Convert SSH keys automatically using WinSCP.com
+
 ## [0.2.0] - 2020-08-05
 
 ### Added