+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Convert SSH keys automatically using WinSCP.com
+- Update dependencies
 
 ## [0.2.0] - 2020-08-05