+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Add support for WinSCP
 - Unix-like aliases (touch, ls, rm)
 - Documentation for installation with PNPM
 - Documentation for installation with Yarn