+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Updated NPM keywords
+
 ## [0.3.0] - 2020-08-11
 
 ### Added
@@ -50,6 +56,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Initial release
 
+[Unreleased]: https://github.com/TheLastZombie/sshpm/compare/0.3.0...HEAD
 [0.3.0]: https://github.com/TheLastZombie/sshpm/releases/tag/0.3.0
 [0.2.0]: https://github.com/TheLastZombie/sshpm/releases/tag/0.2.0
 [0.1.0]: https://github.com/TheLastZombie/sshpm/releases/tag/0.1.0